♻️ Extract standard_runner.go
